requirements.txt
amqp==2.5.2
appdirs==1.4.3
astroid==1.6.6
attrs==19.3.0
Automat==0.8.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
billiard==3.6.1.0
cached-property==1.5.1
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
cliapp==1.20160724
click==6.6
colorama==0.3.7
configparser==4.0.2
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==1.7.1
decorator==4.4.1
defusedxml==0.6.0
Django==1.11.27
djangorestframework==3.9.4
djangorestframework-xml==1.4.0
enum34==1.1.6
fancycompleter==0.8
feedparser==5.1.3
futures==3.3.0
html5lib==0.999999999
httplib2==0.17.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.4.0
incremental==17.5.0
ipaddress==1.0.17
ipython==5.8.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
jasmin==0.9.31
keyring==10.1
keyrings.alt==1.3
kombu==4.6.7
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.4.2
Markdown==2.6.8
mccabe==0.6.1
more-itertools==5.0.0
mutagen==1.36
numpy==1.12.1
PAM==0.4.2
pathlib2==2.3.5
pdbpp==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==4.0.0
prompt-toolkit==1.0.18
ptyprocess==0.6.0
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
Pygments==2.5.2
pygobject==3.22.0
PyHamcrest==1.10.1
pyinotify==0.9.6
pylint==1.9.5
pyOpenSSL==16.2.0
pyparsing==2.4.6
pyrepl==0.9.0
pyserial==3.2.1
python-dateutil==2.8.1
python-smpp==0.1.9
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
quodlibet==3.7.1
redis==3.3.11
requests==2.22.0
requests-toolbelt==0.9.1
scandir==1.10.0
scour==0.32
SecretStorage==2.3.1
service-identity==16.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
smpplib==2.0.1
suds-community==0.8.4
traitlets==4.3.3
ttystatus==0.34
Twisted==19.10.0
txAMQP==0.8.2
urllib3==1.25.7
uTidylib==0.3
vboxapi==1.0
vine==1.3.0
virtualenv==15.1.0
wcwidth==0.1.8
webencodings==0.5
wmctrl==0.3
wrapt==1.11.2
xmltodict==0.12.0
zeep==3.4.0
zipp==1.0.0
zope.interface==4.7.1
